refactor: extract TrendsFullscreenChart from TrendsView (#80, pt 2b)
Moves the ~290-line landscape fullscreen chart out of TrendsView into its own
view, owning its crosshair selection (fullscreenSelectedDate) locally instead
of on TrendsView. Uses the shared TrendsWindowStepper and TrendsAxisScale from
pt 2a. TrendsView's fullScreenCover now presents TrendsFullscreenChart and
passes an onClose closure (replacing the direct showFullscreen mutation).

Verbatim move — no behavior change. TrendsView 1407 → 1131 lines.

Note: two pre-existing Trends UX gaps surfaced during verification (the
fullscreen hero doesn't track the crosshair the way portrait does; the window
stepper crowds the dual-hero in compare mode). Both pre-date this refactor
(confirmed against origin/main) and will be addressed in a separate UX change.

Verified: make lint (0, --strict), make build, make test (272 logic + 13 UI,
all green). Fullscreen verified on simulator (opens, crosshair draws, stepper
and close work).
